Trait minimint_api::config::GenerateConfig

pub trait GenerateConfig: Sized {
    type Params: ?Sized;
    type ClientConfig;

    fn trusted_dealer_gen(
        peers: &[PeerId], 
        max_evil: usize, 
        params: &Self::Params, 
        rng: impl RngCore + CryptoRng
    ) -> (BTreeMap<PeerId, Self>, Self::ClientConfig);
}

Part of a config that needs to be generated to bootstrap a new federation.

Required Associated Types

type Params: ?Sized

type ClientConfig

Required Methods

fn trusted_dealer_gen(
peers: &[PeerId],
max_evil: usize,
params: &Self::Params,
rng: impl RngCore + CryptoRng
) -> (BTreeMap<PeerId, Self>, Self::ClientConfig)

Function that generates the config of all peers locally. This is only meant to be used for testing as the generating machine would be a single point of failure/compromise.